requirements.txt
discord.py
aiohttp
python-dotenv
pytz
